Aligned logic across handlers and added more documentation
