+++ examples/workflow-server/src/common/mcp/workflow-mcp-module.ts
@@ -15,13 +15,13 @@
  ********************************************************************************/
 
 import {
-    CreateNodeMcpToolHandler,
+    CreateNodesMcpToolHandler,
     ElementTypesMcpResourceHandler,
     McpModelSerializer,
     ModifyNodesMcpToolHandler
 } from '@eclipse-glsp/server-mcp';
 import { ContainerModule } from 'inversify';
-import { WorkflowCreateNodeMcpToolHandler } from './workflow-create-node-handler';
+import { WorkflowCreateNodesMcpToolHandler } from './workflow-create-nodes-handler';
 import { WorkflowElementTypesMcpResourceHandler } from './workflow-element-types-handler';
 import { WorkflowMcpModelSerializer } from './workflow-mcp-model-serializer';
 import { WorkflowModifyNodesMcpToolHandler } from './workflow-modify-nodes-handler';
@@ -30,7 +30,7 @@ export function configureWorfklowMcpModule(): ContainerModule {
     return new ContainerModule((bind, unbind, isBound, rebind) => {
         rebind(McpModelSerializer).to(WorkflowMcpModelSerializer).inSingletonScope();
         rebind(ElementTypesMcpResourceHandler).to(WorkflowElementTypesMcpResourceHandler).inSingletonScope();
-        rebind(CreateNodeMcpToolHandler).to(WorkflowCreateNodeMcpToolHandler).inSingletonScope();
+        rebind(CreateNodesMcpToolHandler).to(WorkflowCreateNodesMcpToolHandler).inSingletonScope();
         rebind(ModifyNodesMcpToolHandler).to(WorkflowModifyNodesMcpToolHandler).inSingletonScope();
     });
 }
